Update or install, not both.

+++ setup.sh
@@ -1,10 +1,20 @@
 #!/bin/bash
 mkdir -p /root/.kube
 mkdir -p /opt/atlas
-cd /opt/atlas
-git reset --hard
-git pull
-git clone git@code.penguinzmedia.group:pmg-build-standards/baseline-ansible-setup.git
+mkdir -p /opt/baseline-ansible-setup
+cd /opt/baseline-ansible-setup
+FILE=/opt/baseline-ansible-setup/setup.sh
+if test -f "$FILE"; then
+    echo "$FILE exists, updating."
+    git reset --hard
+    git fetch
+    git pull
+else
+    cd /opt
+    git clone git@code.penguinzmedia.group:pmg-build-standards/baseline-ansible-setup.git
+fi
+
+
 cd /opt/atlas
 cp -f configs/hosts.yaml /etc/ansible/hosts
 cp -f configs/hosts /etc/hosts
